introducing DecimalTypeArg

+++ src/main/java/com/github/pfichtner/pacto/matchers/IntegerTypeArg.java
@@ -4,13 +4,13 @@ import static java.lang.String.format;
 
 public class IntegerTypeArg extends PactoMatcher<String> {
 
-	private final int value;
+	private final long value;
 
-	public IntegerTypeArg(int value) {
+	public IntegerTypeArg(long value) {
 		this.value = value;
 	}
 
-	public int getValue() {
+	public long getValue() {
 		return value;
 	}
